+++ src/Authentication/AuthenticationServiceProvider.php
@@ -5,7 +5,6 @@ namespace BristolSU\Support\Authentication;
 use BristolSU\Support\Authentication\AuthQuery\Generator;
 use BristolSU\Support\Authentication\Contracts\ResourceIdGenerator;
 use Illuminate\Routing\UrlGenerator;
-use BristolSU\Support\Authentication\Contracts\Authentication;
 use Illuminate\Support\ServiceProvider;
 
 class AuthenticationServiceProvider extends ServiceProvider
@@ -14,12 +13,6 @@ class AuthenticationServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind(ResourceIdGenerator::class, AuthenticationResourceIdGenerator::class);
-   
-        $this->app->rebinding('request', function ($app, $request) {
-            $request->setUserResolver(function () use ($app) {
-                return $app->make(Authentication::class)->getUser();
-            });
-        });
     }
 
     public function boot()
@@ -29,10 +22,6 @@ class AuthenticationServiceProvider extends ServiceProvider
         });
         UrlGenerator::macro('getAuthQueryString', function() {
             return app(Generator::class)->getAuthCredentials()->toString();
-        });
-        
-        $this->app['auth']->resolveUsersUsing(function() {
-            return app()->make(Authentication::class)->getUser();
         });
     }