+++ src/SupportServiceProvider.php
@@ -20,6 +20,7 @@ use BristolSU\Support\Permissions\PermissionServiceProvider;
 use BristolSU\Support\Progress\ProgressServiceProvider;
 use BristolSU\Support\Revision\RevisionServiceProvider;
 use BristolSU\Support\Settings\SettingsServiceProvider;
+use BristolSU\Support\Theme\ThemeServiceProvider;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 
@@ -51,7 +52,8 @@ class SupportServiceProvider extends ServiceProvider
         HttpServiceProvider::class,
         RevisionServiceProvider::class,
         SettingsServiceProvider::class,
-        ControlServiceProvider::class
+        ControlServiceProvider::class,
+        ThemeServiceProvider::class
     ];
 
     public function register()